+++ mass_circular_weighing/__init__.py
@@ -2,7 +2,7 @@ from .log import log
 
 __author__ = 'Measurement Standards Laboratory of New Zealand'
 __copyright__ = '\xa9 2024, ' + __author__
-__version__ = '2.0.1'  #.dev0'
+__version__ = '2.0.2.dev0'
 
 from .gui.gui import show_gui
 from .utils.poll_omega_logger import poll_omega_logger